requirements.txt
defusedxml==0.4.1
Django==1.9.5
django-allauth==0.25.2
django-cors-headers==2.2.0
django-filter==1.0.1
djangorestframework==3.5.3
djangorestframework-camel-case==0.2.0
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
Markdown==2.6.8
numpy==1.13.1
oauthlib==1.1.2
openpyxl==2.4.8
pandas==0.22.0
Pillow==3.2.0
psycopg2==2.7.4
PyJWT==1.4.0
python-dateutil==2.7.0
python-social-auth==0.2.1
python3-openid==3.0.10
pytz==2018.3
requests==2.10.0
requests-oauthlib==0.3.1
scikit-learn==0.18.2
six==1.10.0
sklearn==0.0
virtualenv==15.1.0
PyMySQL==0.8.0
MySQL-python==1.2.5